MAke cleanup_files.bash more robust

+++ config/paludis/hooks/ebuild_preinst_pre/cleanup_files.bash
@@ -15,5 +15,5 @@ files=(
 
 einfo "Removing ${files[@]}"
 
-rm -rf "${files[@]}"
+[[ ${#files[@]} -eq 0 ]] || rm -rf "${files[@]}"
